+++ src/main/java/org/example/mirimilibe/global/auth/service/AuthService.java
@@ -1,36 +1,23 @@
 package org.example.mirimilibe.global.auth.service;
 
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Objects;
-import java.util.stream.Stream;
-
 import org.example.mirimilibe.common.Enum.MiliStatus;
 import org.example.mirimilibe.common.Enum.Status;
-import org.example.mirimilibe.common.Enum.TermType;
 import org.example.mirimilibe.global.auth.dto.JwtMemberDetail;
 import org.example.mirimilibe.global.auth.dto.LoginReq;
 import org.example.mirimilibe.global.auth.dto.LoginSuccessRes;
 import org.example.mirimilibe.global.auth.dto.RefreshDTO;
 import org.example.mirimilibe.global.auth.jwt.util.JwtTokenUtil;
 import org.example.mirimilibe.global.error.MemberErrorCode;
-import org.example.mirimilibe.global.error.SmsErrorCode;
 import org.example.mirimilibe.global.error.MilitaryInfoErrorCode;
 import org.example.mirimilibe.global.exception.MiriMiliException;
 import org.example.mirimilibe.member.domain.Member;
-import org.example.mirimilibe.global.auth.dto.SignUpReq;
-import org.example.mirimilibe.member.domain.MemberTerm;
 import org.example.mirimilibe.member.domain.MilitaryInfo;
 import org.example.mirimilibe.member.repository.MemberRepository;
-import org.example.mirimilibe.member.repository.MemberTermRepository;
 import org.example.mirimilibe.member.repository.MilitaryInfoRepository;
-import org.example.mirimilibe.member.service.MemberService;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -42,63 +29,11 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class AuthService {
 	private final MemberRepository memberRepository;
-	private final BCryptPasswordEncoder passwordEncoder;
 	private final AuthenticationManager authenticationManager;
 	private final JwtTokenUtil jwtTokenUtil;
-	private final MemberService memberService;
-	private final MemberTermRepository memberTermRepository;
 	private final MilitaryInfoRepository militaryInfoRepository;
 	private final BlackListService blackListService;
 
-	public void signUp(SignUpReq signUpReq) {
-		//0. 문자 인증 여부 조회
-		/*if( !coolSmsService.isCertificationCompleted(signUpReq.phoneNumber()) ) {
-			throw new MiriMiliException(SmsErrorCode.NEED_SMS_VERIFICATION);
-		}*/
-
-		//1. 약관 동의 검사
-		if (!signUpReq.serviceAgreed() || !signUpReq.privacyPolicyAgreed()) {
-			throw new MiriMiliException(MemberErrorCode.INVALID_MEMBER_PARAMETER);
-		}
-
-		checkDuplicatePhoneNumber(signUpReq.phoneNumber());
-		checkDuplicateNickname(signUpReq.nickname());
-
-		//2. 비밀번호 암호화
-		String encodedPassword = passwordEncoder.encode(signUpReq.password());
-
-		LocalDateTime now = LocalDateTime.now();
-
-		Member member = Member.builder()
-			.number(signUpReq.phoneNumber())
-			.password(encodedPassword)
-			.nickname(signUpReq.nickname())
-			.status(Status.ACTIVE)
-			.createdAt(now)
-			.build();
-
-		//3. 회원 정보 저장
-		memberRepository.save(member);
-
-		//4. 약관 동의 정보 저장
-		List<MemberTerm> memberTerms = Stream.of(
-				TermType.SERVICE,
-				TermType.PRIVACY,
-				signUpReq.marketingConsentAgreed() ? TermType.MARKETING : null)
-			.filter(Objects::nonNull)
-			.map(type -> MemberTerm.builder()
-				.member(member)
-				.termType(type)
-				.agreedAt(now)
-				.build())
-			.toList();
-
-		memberTermRepository.saveAll(memberTerms);
-
-		//5. 군 정보 생성
-		memberService.createMilitaryInfo(signUpReq.miliStatus(), member);
-
-	}
 
 	@Transactional
 	public LoginSuccessRes login(LoginReq loginReq) {
@@ -187,12 +122,6 @@ public class AuthService {
 		blackListService.addBlacklist(accessToken);
 
 		log.info("로그아웃 성공: 사용자 ID={}", memberId);
-	}
-
-	public void checkDuplicatePhoneNumber(String phoneNumber) {
-		if (memberRepository.existsByNumber(phoneNumber)) {
-			throw new MiriMiliException(MemberErrorCode.DUPLICATE_PHONE_NUMBER);
-		}
 	}
 
 	public void checkDuplicateNickname(String nickname) {
